requirements.txt
certifi==2019.6.16
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==4.0.0
pytesseract==0.3.0
Werkzeug==0.15.5
gunicorn==19.9.0
